// frontend/src/generated/types/consumerOrderCreateResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { ConsumerOrderCreateResponseStatus } from './consumerOrderCreateResponseStatus';

export interface ConsumerOrderCreateResponse {
  orderId: string;
  orderNo: string;
  status: ConsumerOrderCreateResponseStatus;
  totalAmount: number;
  orderedAt: string;
}