// frontend/src/generated/types/consumerOrderSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface ConsumerOrderSummary {
  orderId: string;
  orderNo: string;
  status: string;
  totalAmount: number;
  itemCount: number;
  orderedAt: string;
  updatedAt: string;
}